Revert "remove ValueSelectorDisplay"
This reverts commit c9e3d09cadb46312c3ed45e23c2f127a27e03550.
