+++ frontend/src/modules/dashboard/components/ValueInfo.tsx
@@ -14,7 +14,7 @@ import { getAlarmLimitsRequest } from '../../../store/controller/selectors';
 import { setMultiPopupOpen } from '../../app/Service';
 import { AlarmModal } from '../../controllers';
 import { Range } from '../../../store/proto/mcu_pb';
-import { SelectorType } from './constants';
+import { SelectorType, ValueSelectorDisplay } from '../../displays/ValueSelectorDisplay';
 
 const useStyles = makeStyles((theme: Theme) => ({
   rootParent: {
@@ -242,7 +242,6 @@ const ControlValuesDisplay = ({
       : ((alarmLimitsRequest as unknown) as Record<string, Range>)[stateKey];
   const rangeValues = range === undefined ? Range.fromJSON({ lower: '--', upper: '--' }) : range;
   const { lower, upper } = alarmLimits === undefined ? rangeValues : alarmLimits;
-  const value: number | undefined = useSelector(selector);
 
   /**
    * Opens Multistep Popup on Clicking over component
@@ -317,11 +316,7 @@ const ControlValuesDisplay = ({
                   variant="h2"
                   className={`${classes.valueLabel} ${classes.whiteFont}`}
                 >
-                  <React.Fragment>
-                    {value !== undefined && !Number.isNaN(value)
-                      ? value.toFixed(decimal).replace(/^-0$/, '0')
-                      : '--'}
-                  </React.Fragment>
+                  <ValueSelectorDisplay decimal={decimal} selector={selector} />
                 </Typography>
                 {units !== '' && (
                   <Typography
