+++ frontend/src/modules/displays/index.ts
@@ -7,8 +7,10 @@
  * ModeBanner(Current mode text shown in footer of quickstart./set alarms page)
  * ToogleValue(Used in Settings page to toggle values for theme color & unit)
  * ValueDisplay(All the numeric values shown across app uses this component )
+ * ValueSelectorDisplay (Its similar to `ValueDisplay` except that it accepts
  * selector as input & shows the value fetching from redux store using selector )
  */
 export * from './Carousel';
 export * from './ModeBanner';
 export * from './ValueDisplay';
+export * from './ValueSelectorDisplay';
